refactor: move useHigherUSDValue hook to fiatAmount module
