+++ apps/cowswap-frontend/src/modules/account/containers/AccountDetails/SurplusCard.tsx
@@ -2,8 +2,9 @@ import { transparentize } from 'polished'
 import styled from 'styled-components/macro'
 
 import QuestionHelper, { QuestionWrapper } from 'legacy/components/QuestionHelper'
-import { useHigherUSDValue } from 'legacy/hooks/useStablecoinPrice'
 import { ExternalLink } from 'legacy/theme'
+
+import { useHigherUSDValue } from 'modules/fiatAmount'
 
 import { FiatAmount } from 'common/pure/FiatAmount'
 import { TokenAmount } from 'common/pure/TokenAmount'
