+++ apps/cowswap-frontend/src/modules/fiatAmount/hooks/useHigherUSDValue.ts
