+++ apps/cowswap-frontend/src/modules/fiatAmount/index.ts
@@ -1,0 +1,1 @@
+export { useHigherUSDValue, useCoingeckoUsdValue } from './hooks/useHigherUSDValue'
