+++ apps/cowswap-frontend/src/modules/swap/containers/TradeSummary/index.tsx
@@ -1,10 +1,11 @@
 import { Percent } from '@uniswap/sdk-core'
 
-import { useHigherUSDValue } from 'legacy/hooks/useStablecoinPrice'
 import TradeGp from 'legacy/state/swap/TradeGp'
 
 import { TradeSummaryContent } from 'modules/swap/pure/TradeSummary'
 import { useWalletDetails } from 'modules/wallet'
+
+import { useHigherUSDValue } from '../../../fiatAmount/hooks/useHigherUSDValue'
 
 // Sub-components
